requirements.txt
flask==2.3.3
werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
gunicorn==21.2.0
redis==5.0.0